requirements.txt
numpy
opencv-contrib-python
PyYAML
pytest